+++ components/StoriesAvatar.tsx
